The class file has been renamed, and the class instance has been moved out of the index and into its own file. Updated index file accordingly
